Editing
I've done quite a big edit, but most of what I changed is the order. I started with the sections people are likely to care about the very first time they use mbed CLI, and moved to the end the sections that will take a couple of days to reach (like updating libraries). I also put the sections in the order we'd probably like to encourage people to work at, so for example testing comes before publishing.

I’m assuming you can generate the ToC automatically. 

I had a few questions:
1 .Why do I have to install both Git and Mercurial? Why can’t I pick one?
2. For the status check before publishing, what do M and ?? mean, and are there other options?
3. Why do we tell people to manually edit the path to the compilers if we have `ARM_PATH`, `GCC_ARM_PATH`, `IAR_PATH` - defines the default path to ARM Compiler, GCC ARM and IAR Workbench toolchains. Default: none.?