Add more options for configuring compilers
We added a few more options for configuring compilers in the mbed-os
tools, so I'm adding those options to the readme here. The new options
are:
 - Setting env variables
 - `PATH` detection

Further, I updated the order of the subsections to reflect their
priority. Things earlier in the list configure the compiler in preference
to things later in the list.

Finally, there is one edit for consistency. Previously we had a command
line `blah --global blah blah` and in the text it referred to `-G`. I
changed the command line to be consistent