Parse environment markers when auto installing.
The following change in Mbed OS added an environment marker to
the python requirements file:
https://github.com/ARMmbed/mbed-os/pull/10096/files#diff-b4ef698db8ca845e5845c4618278f29aR24
Mbed CLI did not know how to parse this, so a warning was always
generated. This change now parses the "platform_system" marker, and
adds a mechanism to add more if needed in the future.